+++ sensors/LightIntensityReading.h
@@ -32,7 +32,7 @@ public:
     void printValues(Print & out, NewLiner & newLine) const;
 
     virtual void registerSensorValues(SensorReading ** valueArray);
-    virtual void printValue(int valueId, Print & out);
+    virtual void printValue(uint8_t valueId, Print & out);
 
 protected:
     const __FlashStringHelper * getErrText() const;
